+++ crates/cli/tests/run_typescript_test.rs
@@ -16,9 +16,9 @@ where
 
     let sandbox = create_sandbox_with_config(
         "typescript",
-        Some(&workspace_config),
-        Some(&toolchain_config),
-        Some(&tasks_config),
+        Some(workspace_config),
+        Some(toolchain_config),
+        Some(tasks_config),
     );
 
     sandbox.enable_git();