+++ crates/core/config/src/toolchain/rust.rs
@@ -12,7 +12,7 @@ fn validate_rust_version(value: &str) -> Result<(), ValidationError> {
 #[serde(default, deny_unknown_fields, rename_all = "camelCase")]
 pub struct RustConfig {
     #[serde(skip_serializing_if = "is_default")]
-    pub cargo_bins: Vec<String>,
+    pub bins: Vec<String>,
 
     #[serde(skip_serializing_if = "is_default")]
     pub sync_toolchain_config: bool,