+++ crates/node/platform/tests/project_aliases_test.rs
@@ -11,9 +11,9 @@ async fn get_aliases_graph() -> (ProjectGraph, Sandbox) {
 
     let sandbox = create_sandbox_with_config(
         "project-graph/aliases",
-        Some(&workspace_config),
-        Some(&toolchain_config),
-        Some(&tasks_config),
+        Some(workspace_config),
+        Some(toolchain_config),
+        Some(tasks_config),
     );
 
     let mut workspace = load_workspace_from(sandbox.path()).await.unwrap();