+++ crates/rust/lang/src/cargo_toml.rs
@@ -3,6 +3,7 @@ use cached::proc_macro::cached;
 use cargo_toml::Manifest as CargoToml;
 use moon_error::MoonError;
 use moon_lang::config_cache_container;
+use starbase_utils::glob;
 use std::path::{Path, PathBuf};
 
 pub use cargo_toml::*;
@@ -15,6 +16,7 @@ config_cache_container!(CargoTomlCache, CargoToml, CARGO.manifest, read_manifest
 
 pub trait CargoTomlExt {
     fn get_detailed_workspace_dependency(&self, name: &str) -> Option<DependencyDetail>;
+    fn get_member_manifest_paths(&self, root_dir: &Path) -> Result<Vec<PathBuf>, MoonError>;
 }
 
 impl CargoTomlExt for CargoToml {
@@ -35,5 +37,29 @@ impl CargoTomlExt for CargoToml {
             },
             Dependency::Detailed(detail) => detail.to_owned(),
         })
+    }
+
+    fn get_member_manifest_paths(&self, root_dir: &Path) -> Result<Vec<PathBuf>, MoonError> {
+        let mut paths = vec![];
+
+        let Some(workspace) = &self.workspace else {
+            return Ok(paths);
+        };
+
+        let mut patterns = workspace.members.clone();
+
+        for exclude in &workspace.exclude {
+            patterns.push(format!("!{exclude}"));
+        }
+
+        for manifest_dir in glob::walk(root_dir, &patterns)? {
+            let manifest_path = manifest_dir.join(CARGO.manifest);
+
+            if manifest_path.exists() {
+                paths.push(manifest_path);
+            }
+        }
+
+        Ok(paths)
     }
 }